updated bkz for scholar author search
just added in the an author for the scholar search. This enables slightly better searches. Still should test further because I noticed issues when attempting to find low citations for scholarly articles. I believe this is because a large amount of the data comes from goodreads and it should probably be crosschecked with amazon. Will work on that later